Code refactoring (part 4). Fixing string concatination and array copy

+++ src/main/java/com/pony/database/CellInputStream.java
@@ -145,8 +145,8 @@ final class CellInputStream implements CellInput {
         int ch2 = read();
         int ch3 = read();
         int ch4 = read();
-        return (int) ((ch1 << 24) + (ch2 << 16) +
-                (ch3 << 8) + (ch4 << 0));
+        return (ch1 << 24) + (ch2 << 16) +
+                (ch3 << 8) + (ch4 << 0);
     }
 
     public long readLong() throws IOException {
